Corrige conflitos e mudanças de pacotes para a pull request #135
