+++ src/main/java/br/ufg/inf/sempreufg/dao/BackupBancoDao.java
@@ -1,4 +1,4 @@
-package br.ufg.inf.dao;
+package br.ufg.inf.sempreufg.dao;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
@@ -7,23 +7,23 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 
-import br.ufg.inf.enums.MedidasTempo;
-import br.ufg.inf.modelo.ParametrosBackupBanco;
+import br.ufg.inf.sempreufg.enums.MedidasTempo;
+import br.ufg.inf.sempreufg.modelo.ParametrosBackupBanco;
 
 public class BackupBancoDao {
-	
+
 	private Connection con;
-	
+
 	public BackupBancoDao(Connection con) {
 		super();
 		this.con = con;
 	}
-	
+
 	public void salvarParametrosBackupBanco(ParametrosBackupBanco parametros) {
 		StringBuilder sql = new StringBuilder();
 		sql.append("insert into parametros_backup_banco (diretorio_backup, diretorio_padrao_execucao_backup, tempo, unidade) ");
 		sql.append("values(?, ?, ?, ?)");
-		
+
 		PreparedStatement preparedStatement;
 		try {
 			preparedStatement 	= getCon().prepareStatement(sql.toString());
@@ -33,56 +33,56 @@ public class BackupBancoDao {
 			preparedStatement.setString(4, parametros.getUnidade().toString());
 			preparedStatement.execute();
 			preparedStatement.close();
-		 
+
 		} catch (SQLException e) {
 			e.printStackTrace();
 		}
 	}
-	
+
 	public void removerParametros(){
 		StringBuilder sql = new StringBuilder();
 		sql.append("delete from parametros_backup_banco");
-		
+
 		PreparedStatement preparedStatement;
 		try {
 			preparedStatement 	= getCon().prepareStatement(sql.toString());
 			preparedStatement.execute();
 			preparedStatement.close();
-		 
+
 		} catch (SQLException e) {
 			e.printStackTrace();
 		}
 	}
 
 	public List<ParametrosBackupBanco> consultarParametrosBackupBanco() {
-		List<ParametrosBackupBanco> parametros = new ArrayList<ParametrosBackupBanco>(); 
-		
+		List<ParametrosBackupBanco> parametros = new ArrayList<ParametrosBackupBanco>();
+
 		StringBuilder sql = new StringBuilder();
 		sql.append("select diretorio_backup, 				 ")
 		   .append("       diretorio_padrao_execucao_backup, ")
 		   .append("       tempo, 							 ")
 		   .append("       unidade 							 ")
 		   .append("  from parametros_backup_banco			 ");
-		
+
 		PreparedStatement preparedStatement;
 		try {
 			preparedStatement 	= getCon().prepareStatement(sql.toString());
 			ResultSet resultSet = preparedStatement.executeQuery();
-		 
+
 			while(resultSet.next()){
 				ParametrosBackupBanco parametro = new ParametrosBackupBanco();
 				parametro.setDiretorioBackup(resultSet.getString("diretorio_backup"));
 				parametro.setDiretorioPadraoExecucaoBackup(resultSet.getString("diretorio_padrao_execucao_backup"));
 				parametro.setTempo(resultSet.getInt("tempo"));
 				parametro.setUnidade(MedidasTempo.valueOf(resultSet.getString("unidade")));
-				
+
 				parametros.add(parametro);
 			}
-			
+
 		} catch (SQLException e) {
 			e.printStackTrace();
 		}
-			 
+
 		return parametros;
 	}
