+++ src/main/java/br/ufg/inf/sempreufg/enums/TipoBanco.java
@@ -1,4 +1,4 @@
-package br.ufg.inf.enums;
+package br.ufg.inf.sempreufg.enums;
 
 public enum TipoBanco {
 	MARIA_DB, POSTEGRESQL;
