+++ src/main/java/br/ufg/inf/sempreufg/factory/ConnectionFactoryPostegres.java
@@ -1,22 +1,23 @@
-package br.ufg.inf.factory;
+package br.ufg.inf.sempreufg.factory;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.SQLException;
 
-public class ConnectionFactoryMariaDb {
+public class ConnectionFactoryPostegres {
+
 	private Connection con;
-	
+
 	public Connection getConnection() throws ClassNotFoundException, SQLException{
 		if (con == null) {
 			criarConexao();
 		}
-		
+
 		return con;
 	}
-	
+
 	private void criarConexao() throws SQLException, ClassNotFoundException{
-		Class.forName("org.mariadb.jdbc.Driver");
-		this.con = DriverManager.getConnection("jdbc:mariadb://localhost:3306/sempre_ufg", "root", "senha");
+		Class.forName("org.postgresql.Driver");
+		this.con = DriverManager.getConnection("jdbc:postgresql://localhost:5432/sempre_ufg", "postgres", "duckmaster");
 	}
 }
