+++ src/main/java/br/ufg/inf/sempreufg/mb/StartMB.java
@@ -1,4 +1,4 @@
-package br.ufg.inf.mb;
+package br.ufg.inf.sempreufg.mb;
 
 import javax.faces.application.FacesMessage;
 import javax.faces.bean.ManagedBean;
@@ -8,9 +8,9 @@ import javax.faces.context.FacesContext;
 @ManagedBean(name="StartMB")
 @ViewScoped
 public class StartMB {
-	
+
 	public void start(){
 			FacesContext.getCurrentInstance().addMessage("Sucesso", new FacesMessage("Viu Nao acontece nada"));
 	}
-	
+
 }
